Fix template expand with nested %if within an %else condition
The template expander is very basic for the %if/%else/%end condition
support and nested %if is not possible.  However, we can almost use
a nested %if after a %else and this commit fixes the support for that.
We can now write nested conditions:

%if error
 ...
%else
%if debug
 ...
%end

Note: a second '%end' is not necessary because this closes all
opened conditions.